RL/requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
bitsandbytes==0.41.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.1
colorama==0.4.6
DataProperty==1.0.1
datasets==2.14.4
dill==0.3.7
docker-pycreds==0.4.0
einop==0.0.1
einops==0.6.1
faiss-cpu==1.7.4
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.37
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
jsonlines==3.1.0
lit==16.0.6
MarkupSafe==2.1.3
mbstrdecoder==1.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
openai==0.27.8
packaging==23.1
pandas==1.5.3
pathtools==0.1.2
pathvalidate==3.1.0
peft==0.4.0
portalocker==2.7.0
protobuf==4.24.3
psutil==5.9.5
pyarrow==12.0.1
pybind11==2.11.1
pycountry==22.3.5
pytablewriter==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
rouge-score==0.1.2
sacrebleu==1.5.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.1
sqlitedict==2.1.0
sympy==1.12
tabledata==1.3.1
tcolorpy==0.1.3
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
tqdm-multiprocess==0.0.11
transformers==4.31.0
triton==2.0.0
typepy==1.3.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wandb==0.15.11
xxhash==3.3.0
yarl==1.9.2
zstandard==0.21.0
